requirements.txt
langchain
langchain-openai
python-dotenv
django
djangorestframework
langchain_community
redis